Make initial theming
Some colors might change, but the idea is to make it easy to change the
colors of the app by changing the `assets/theme/index.js` file.

Also, a bunch of premade styles for components that will probably be
reused, especially by the articles which should mostly be independent of
the source code.