Allow for including fonts as assets
To allow for changing fonts in the theme file without touching the
source code or the `index.html` file.

Right now, downloading fonts would require adding them to the css by
`@font-face`, and that's non trivial for non-techies. The easiest way I
can think of right now would be to have them choose the fonts from
Google Fonts and copy paste the embedded css code in a file, and that's
what I'm doing right now, as well as adding detailed instructions in the
README.md file.