Install jest and make the first test
The first test is for `src/build/articles.js`. Had to restructure the
file and the way the build process is made in order to make the file
more test friendly, as well as make it cleaner for project itself. Now,
instead of running the script directly in the build step, we export the
functions and import them in `bin/build-articles`, a binary that we run
for the build step.