Overhaul tests and move some files around
We're changing the tests because jest has trouble mocking ESModules and
browsers can't work with CommonJS. Ideally, you can use both if you're
using a compiler like a webpack, which we're not.

So now we're rethinking tests to not use mocks (which is probably a good
thing). Instead of doing unit tests for every function on each file
leaving each test file side by side with its subject, we're having a
`test` directory holding all tests, and they're more general and working
with the whole DOM directly. They're like integration tests, but a
little more low level, yet not so low level like your usual unit tests.

There might be some exceptions to the "no mocks" rule, like the build
script test that mocks Node's `fs` module so we don't need to depend on
making special directories to test the build script on, which in the end
amounts to the same thing as mocking them. Jest has no trouble mocking
these modules, it seems like it's only our own written modules that are
troublesome.

Other than that, we're changing `content/articles.js` to a named export
instead of a default one to standardize named exports to all files, and
we're moving `navigation/set-links.js` to `components/anchor.js` because
we should eventually make custom components, and since we're already
customizing the default behavior of anchors, we should treat them as a
component (which we've already been doing).