Improve built articles
Currently, they're an array of objects with a name and a content key.
Obviously, they should have been just an object with every key being its
name and having the content as its value.

This way, we don't need to iterate through the array looking for the
name of the article, you just access it directly through the hashmap.