Implement project settings
A JSON file for settings, to make it easier to change very specific
things such as labels and defaults that should vary from fork to fork.

The inspiration behind this feature came because I wanted to implement a
default map, so that if you opened that link with no map search param it
would load a default map always, and whenever you input an invalid map
or an empty map, it would send you to this map and erase the search
param. It's obvious that such default map needed a way to change it more
easily, especially for non technical DMs forking this. I was also
already thinking about localization on the table of contents label, and
now with the "close" label on the top of the articles container, this
settings feature should also take care of it.