Implement the initDom function
This function initializes the DOM with the actual `index.html` file
used by the project, meant to be used on tests that have to initialize
a DOM with enough for the tests to run.

With this, we will have less code repeating and we won't need to keep
changing all these tests every time we add new element with an id that
will be read by some script on `main.js`, everything gets cleaner.

Particularly, the last few PRs where we implemented maps and styled the
articles had new elements being added and thus forced us to update a
bunch of unrelated tests because some getElementById failed.