Implement browser back function
Because of the way we navigate without reloading the page but pushing to
history, pressing the back button causes the URL to return to what it
was previously but the page wouldn't update because we wouldn't call the
detect functions.

This commit fixes it with the `popstate` event. We also improved and
cleaned up a little more of the previous code revolving navigation, like
the old `toArticle()` and `toMap()` functions, that were created mostly
to allow buttons to navigate but in the end only using anchors to
navigate because it makes more sense on a web app, also because when you
hover the mouse on an anchor it shows where it's going to take you while
buttons can't do that cleanly. So if we're just using anchors for that,
there's no need to have these functions existing.