Fix anchor hrefs
Anchors with `""` on `tomap` or `toarticle` has broken hrefs. They
worked and sent you to the right place though.

Also changes the way `tomap` works when sending you to the default map,
now an anchor to the default map just deletes the `map` search param in
the same way a link with `tomap=""` would.