Style tables for articles
We might need to put tables on articles. This commit not only styles it
but also reorganizes the stylesheet to make better use of tailwindcss
layers, which had a lot of reduntant wrong stuff previously. Also
includes float classes that might be reusable for images and other
stuff.

It also adds `assets/articles` to tailwindcss's build content, as
without it it wouldn't scan articles and wouldn't include classes in
there, which would cause custom classes not to work if they only showed
up on articles.