Fix anchors not erasing URL hash
If you clicked a link on a table of contents, it would set a hash on the
current location. Then, if you clicked a link to another article, map or
the close location link, it would do its job but the hash would remain
there.

This commit fixes this behaviour by making all links always clean
hashes.