Add top bar
Adds a top bar to add as a menu of sorts. Also transfers the dark/light
theme switcher link to it, as it would make more sense there.

As things got rearranged, we also switched around some divs for more
semantic tags like main for the main container and article for the
article container.