Parse article titles on build
To make a proper index page, we need to print article titles. Using the
file names won't work very well for a good experience, so I decided to
parse it out of the `<h1>` tag on the article. Now all articles require
a h1 element, which they all would probably have anyway.