Add index links on top bar
Add links to the articles index and the maps index on the top bar. To do
this, we needed to add the `tomenu` attribute functionality to anchors
in a similar manner to the previous `tomap` and `toarticle` attributes.
This new one has special behavior with the `toarticle` attribute, as
both menus and articles occupy the same container, only one of them can
be loaded at the same time, so I'm giving priority to the `tomenu`
property, which means an anchor with both `tomenu` and `toarticle` will
ignore the `toarticle`, and if there is either a menu or article loaded
and you click an anchor to the other, the previous one will get cleared
out of the search params.

There's also a couple of typo corrections in this commit as well, on the
anchor tests.