Fix navigation between articles and menus bug
There was a bug due to how going from an article to a menu and vice
versa wouldn't clear the `current` variable in either, causing the
detect function to not do anything when going from an article back to
the last menu loaded or going from a menu to the previously loaded
article.

This commit just removes this variable, it is no long necessary. It
never was in fact necessary, we just wanted to prevent any changes from
happening when the user went from an article to the same article (would
probably cause the article to scroll back to the top), which might be
worse UX when you think about it, as nothing happening when the user
clicks a link/button should be avoided.