Refactor set anchors
The plan is to fetch for the maps, articles and menu files instead of
loading them all into memory and getting them when navigating. This
means we have to fetch when the user clicks a link, which triggers a
rerender which triggers the detects. Since each detect will fetch,
they're all async now, meaning the `setAnchors()` call at the end of
`render()` will hit nothing.

This commit refactors `setAnchors()` to receive the list of anchors it's
setting up, so we can call it directly from the detect functions after
fetching is finished, giving it only the new anchors instead of the
whole app.