Refactor detect article with fetch
Instead of loading from `window.imports.articles`, we are fetching
directly into `/assets/articles` for the asked article. This should free
up memory as we're won't be loading all articles onto memory.

We decided for this change because we were already fetching data, as
articles and maps would already fetch for images when they had them so
it already wasn't 100% free of load times. Secondly, this change
simplifies the code, makes the app less memory intensive and frees up
the need for building articles. It also allows for better error
handling, as instead of just showing 404 messages we can end up showing
everything fetch returns.

This change isn't done, as we need to do the same to maps, menus and
rethink the build step, as right now it's still building articles and
loading them to memory.