Refactors build settings
On deploy, all fetch article calls were failing while on dev they were
not. This was because `/assets/articles/<some-article>.html` was being
concatenated to `https://<username>.github.io` instead of
`https://<username>.github.io/<repo-name>`. To fix it, we needed to add
the repo name on to the fetch URI only on the deploy.

To do this, we're including the articles URL path on the built settings.
We can add the repo name on a env var through the github action that
calls it, and leaving that env var empty would cause no changes to de
dev environment.