Fix flickering and improve tests
Whenever we navigated anywhere, the screen would flicker. That was
because the map was reloading even though we didn't changed it. Changing
just the map would cause the menu or article to flicker as well, for the
same reason.

To fix it, we returned the code that was preventing needless re-renders
with a `current` variable storing the currently loaded article/menu/map
and not doing anything in case the new one to be render is the same.

This code was removed because it caused a bug when navigating from an
article to a menu and back to the same article, or from a menu to an
article and back to the same menu. So we put this code back but in a way
that would not cause this bug, as well as with a test to ensure this bug
hasn't returned

While making the new tests (the one for the old bug and for the new anti
flickering feature), we saw some tests with typos and one test was
missing (anchors to both menus and articles), so this commit also fixes
it.