Update content, docs and style
In preparation for a launch, I'm removing all of colors, fonts and icons
related to my future project, which should be a fork of this one.

I also revamped all the maps and articles in order to turn them into
more documentation, as well as update anything that wasn't up to date on
the .md files in this project as well. Doing so, I went away with
splitting primary colors on theme between dark and light theme, and also
put a way on the project settings to let us set the favicon.