Implements search data build step
Adds a script to build a json file for the search data. The script
parses all maps and articles and makes a hashtable with their titles as
keys and tables with `tomap` and `toarticle` keys as their values,
saving it as json.

The idea is that we fetch this on load, so whenever the user types
something in the search bar we can filter whatever they're typing and
get a link to it. If there's both a map and an article with that name,
the link should go to both.