Disable cache when fetching for articles, maps and menus
Because they are constantly changing and I can't control the GitHub
Pages server headers (not that I know), it's easier to just disable the
cache than to expect your users to know how and when to refresh them.

We could make an more robust build system to include some versioning on
the names of every fetchable resource but I don't think it's going to be
a problem for now as HTML and JSON files are very light, so fetching
them every time shouldn't be problematic.